requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
pydantic==2.12.5
pydantic-settings==2.12.0
Jinja2==3.1.6
apscheduler==3.11.1
acme==5.2.1
josepy==2.2.0
cryptography==46.0.3
azure-identity==1.25.1
azure-keyvault-certificates==4.10.0
azure-mgmt-dns==9.0.0
loguru==0.7.3
aiohttp==3.13.2
